refactor(sea): extract context and metabolism operations
